+++ src/RobotInterface/include/WalkingControllers/RobotInterface/Helper.h
@@ -20,6 +20,7 @@
 #include <yarp/dev/IPositionControl.h>
 #include <yarp/dev/IPositionDirect.h>
 #include <yarp/dev/IVelocityControl.h>
+#include <yarp/dev/IInteractionMode.h>
 #include <yarp/sig/Vector.h>
 #include <yarp/os/Timer.h>
 
@@ -44,6 +45,7 @@ namespace WalkingControllers
         yarp::dev::IVelocityControl *m_velocityInterface{nullptr}; /**< Position control interface. */
         yarp::dev::IControlMode *m_controlModeInterface{nullptr}; /**< Control mode interface. */
         yarp::dev::IControlLimits *m_limitsInterface{nullptr}; /**< Encorders interface. */
+        yarp::dev::IInteractionMode *m_InteractionInterface{nullptr}; /**< Stiff/compliant mode interface. */
 
         std::unique_ptr<WalkingPIDHandler> m_PIDHandler; /**< Pointer to the PID handler object. */
 
@@ -62,7 +64,9 @@ namespace WalkingControllers
         iDynTree::VectorDynSize m_jointVelocitiesBounds; /**< Joint Velocity bounds [rad/s]. */
         iDynTree::VectorDynSize m_jointPositionsUpperBounds; /**< Joint Position upper bound [rad]. */
         iDynTree::VectorDynSize m_jointPositionsLowerBounds; /**< Joint Position lower bound [rad]. */
-
+         std::vector<yarp::dev::InteractionModeEnum> m_isJointModeStiffVector;/**< Joint is in the stiff or compliance mode */
+         std::vector<yarp::dev::InteractionModeEnum> m_JointModeStiffVectorDefult;/**< All the joints are in the stiff  mode */
+         std::vector<yarp::dev::InteractionModeEnum> m_currentModeofJoints;/**< Joint is in the stiff or compliance mode based on the walking architecture phases */
         // yarp::sig::Vector m_positionFeedbackDegFiltered;
         yarp::sig::Vector m_velocityFeedbackDegFiltered; /**< Vector containing the filtered joint velocity [deg/s]. */
         std::unique_ptr<iCub::ctrl::FirstOrderLowPassFilter> m_positionFilter; /**< Joint position low pass filter .*/
@@ -81,6 +85,7 @@ namespace WalkingControllers
         std::unique_ptr<iCub::ctrl::FirstOrderLowPassFilter> m_rightWrenchFilter; /**< Right wrench low pass filter.*/
         bool m_useWrenchFilter; /**< True if the wrench filter is used. */
 
+        std::vector<bool>  m_jointModes; /**< True if the joint is in the stiff mode */
         double m_startingPositionControlTime;
         bool m_positionMoveSkipped;
 
@@ -94,7 +99,7 @@ namespace WalkingControllers
          * @return true in case of success and false otherwise.
          */
         bool getWorstError(const iDynTree::VectorDynSize& desiredJointPositionsRad,
-                           std::pair<std::string, double>& worstError);
+                           std::pair<int, double> &worstError);
 
         /**
          * Switch the control mode.
@@ -206,6 +211,12 @@ namespace WalkingControllers
         int getActuatedDoFs();
 
         WalkingPIDHandler& getPIDHandler();
+
+        /**
+         * Set the intraction mode of the joints(stiff/compliant).
+         * @return true in case of success and false otherwise.
+         */
+        bool setInteractionMode();
     };
 };
 #endif